jellylorum/requirements.txt
Django==1.3
South==0.7.3
Whoosh==2.4.1
coverage==3.5.1b1
distribute==0.6.19
django-haystack==2.0.0
django-nose==0.1.3
lxml==2.3
nose==1.1.2
pyquery==1.0
wsgiref==0.1.2
yolk==0.4.1
